+++ R/cache.R
@@ -684,8 +684,10 @@ setMethod(
           }
           # Now changed objects
           if (length(unlist(changedModEnvObjs))) {
+            # if (identical(currentModule(object), "canClimateData")) browser()
             Map(nam = names(changedModEnvObjs), objs = changedModEnvObjs, function(nam, objs) {
-              objNames <- names(objs$.objects)
+              objNames <- names(objs$.objects) # used to be "names(...)" -- but don't want `._` objs
+              objNames <- grep("^._.+", objNames, value = TRUE, invert = TRUE)
               if (!is.null(objNames))
                 list2env(mget(objNames, envir = simFromCache@.xData$.mods[[nam]][[".objects"]]),
                          envir = simPost@.xData$.mods[[nam]][[".objects"]])
@@ -927,16 +929,21 @@ if (!exists("objSize")) {
 #' utils::object.size(a)
 objSize.simList <- function(x, quick = TRUE, ...) {
 
-  total <- obj_size(x, quick = TRUE)
-  aa <- objSize(x@.xData, quick = quick, ...)
+  total <- try(obj_size(x, quick = TRUE), silent = TRUE) # failing due to lobstr issue #72
+  if (!is(total, "try-error")) {
+    aa <- objSize(x@.xData, quick = quick, ...)
 
-  simSlots <- grep("^\\.envir$|^\\.xData$", slotNames(x), value = TRUE, invert = TRUE)
-  names(simSlots) <- simSlots
-  otherParts <- objSize(lapply(simSlots, function(slotNam) slot(x, slotNam)), quick = quick, ...)
+    simSlots <- grep("^\\.envir$|^\\.xData$", slotNames(x), value = TRUE, invert = TRUE)
+    names(simSlots) <- simSlots
+    otherParts <- objSize(lapply(simSlots, function(slotNam) slot(x, slotNam)), quick = quick, ...)
 
-  if (!quick)
-    attr(total, "objSizes") <- list(sim = attr(aa, "objSize"),
-                                    other = attr(otherParts, "objSize"))
+    if (!quick)
+      attr(total, "objSizes") <- list(sim = attr(aa, "objSize"),
+                                      other = attr(otherParts, "objSize"))
+
+  } else {
+    total <- NA
+  }
 
   return(total)
 }