+++ R/simulation-simInit.R
@@ -576,7 +576,6 @@ setMethod(
 
       ## do multi-pass if there are parent modules; first for parents, then for children
       all_parsed <- FALSE
-      # browser(expr = exists("._simInit_5"))
       while (!all_parsed) {
         sim <- .parseModule(sim,
                             as.list(sim@modules),
@@ -1332,6 +1331,9 @@ simInitAndSpades <- function(times, params, modules, objects, paths, inputs, out
             runFnCallAsExpr <- is.null(attr(sim, "runFnCallAsExpr"))
           }
           if (runFnCallAsExpr) {
+            pkgs <- Require::extractPkgName(unlist(moduleMetadata(sim, currentModule(sim))$reqdPkgs))
+            pkgs <- c(pkgs, "stats")
+            do.call(box::use, lapply(pkgs, as.name))
             sim <- Cache(.inputObjects, sim,
                          .objects = objectsToEvaluateForCaching,
                          notOlderThan = notOlderThan,
@@ -1547,13 +1549,24 @@ loadPkgs <- function(reqdPkgs) {
     # Check for SpaDES.core minimum version
     checkSpaDES.coreMinVersion(allPkgs)
     allPkgs <- grep("^SpaDES.core\\>", allPkgs, value = TRUE, invert = TRUE)
+
+    pkgsDontLoad <- getOption("spades.reqdPkgsDontLoad", NULL)
+    allPkgs <- reqdPkgsDontLoad(allPkgs, pkgsDontLoad)
+
     if (getOption("spades.useRequire")) {
       getCRANrepos(ind = 1) # running this first is neutral if it is set
-      Require(allPkgs, standAlone = FALSE, upgrade = FALSE)
+      Require(allPkgs, require = require, standAlone = FALSE, upgrade = FALSE)
+      if (!is.null(pkgsDontLoad)) {
+        verbose <- getOption("reproducible.verbose")
+        Require::Require(pkgsDontLoad, require = FALSE, standAlone = FALSE,
+                         upgrade = FALSE, verbose = verbose - 1)
+      }
       # RequireWithHandling(allPkgs, standAlone = FALSE, upgrade = FALSE)
     } else {
-      allPkgs <- unique(Require::extractPkgName(allPkgs))
-      loadedPkgs <- lapply(allPkgs, require, character.only = TRUE)
+      if (!getOption("spades.useBox")) {
+        allPkgs <- unique(Require::extractPkgName(allPkgs))
+        loadedPkgs <- lapply(allPkgs, base::require, character.only = TRUE)
+      }
     }
   }
 }
@@ -1905,3 +1918,6 @@ simNestingOverride <- function(sim, mBase) {
   sim[["._simNesting"]][len] <- paste0(modName8Chars, ":", cli::col_green(sim@current$eventType))
   sim[["._simNesting"]]
 }
+
+isMacOSX <- function()
+  isMac <- tolower(Sys.info()["sysname"]) == "darwin"